backend/requirements.txt
django==4.2.10
openai==1.12.0
djangorestframework==3.14.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1 